refactor: 3c (commands) — extract cmd_apply + shared install_common
Move the apply cluster (cmd_apply, _cmd_apply_detect_dry_run,
_apply_variant_install, _apply_legacy_install, features-dir helpers) to
cli/cmd_apply.py. Extract the three helpers shared by apply and upgrade —
exclude_for_level, copy_governed_or_shared, post_install_finalize — into
cli/install_common.py as public API; cmd_upgrade (still in govkit) now imports
them from there. govkit re-exports cmd_apply for main(); dropped the now-unused
stack_select trio and TYPE_CHECKING/RepoProfile imports, kept resolve_options as
a test re-export. govkit.py: 740 -> 504 lines. No behavior change; 807 green.
