+++ cli/govkit.py
@@ -30,12 +30,9 @@ import argparse
 import shutil
 import sys
 from pathlib import Path
-from typing import TYPE_CHECKING
-
-if TYPE_CHECKING:
-    from .detect import RepoProfile
 
 from . import paths, version
+from .cmd_apply import cmd_apply
 from .cmd_init import _resolve_starter_dir as _resolve_starter_dir
 from .cmd_init import cmd_init
 from .cmd_list import cmd_list
@@ -43,7 +40,13 @@ from .cmd_stack import cmd_stack_apply, cmd_stack_list
 from .cmd_validate import cmd_validate
 from .fs import copy_entry
 from .fs import is_user_edited as is_user_edited
-from .manifest import load_manifest, resolve_options, resolve_variant_files
+from .install_common import (
+    copy_governed_or_shared,
+    exclude_for_level,
+    post_install_finalize,
+)
+from .manifest import load_manifest, resolve_variant_files
+from .manifest import resolve_options as resolve_options
 from .marker import (
     _compare_version,
     read_govkit_marker,
@@ -59,249 +62,10 @@ from .marker import (
     _reset_shape_migration_warning as _reset_shape_migration_warning,
 )
 from .marker import read_govkit_level as read_govkit_level
-from .stack_select import (
-    apply_stack_overlay,
-    print_detection_summary,
-    resolve_stack_choice,
-)
-
-# PR 6c: L5-only architecture docs. These live in docs/backend/architecture/
-# alongside the universal baseline files for repo-self-governance purposes,
-# but they must NOT install at L3/L4 — doctor D007 (LLM-leakage in non-L5)
-# is the canary. Future PR may relocate these into a separate dir and drop
-# the exclusion machinery.
-_L5_ONLY_GOVERNED_BASENAMES: set[str] = {
-    "AGENT_ARCHITECTURE.md",
-    "LLM_GATEWAY_CONTRACT.md",
-    "GUARDRAILS_CONTRACT.md",
-    "OBSERVABILITY_LLM_CONTRACT.md",
-    "EVALUATION_LLM_CONTRACT.md",
-}
-
-
-def _exclude_for_level(level: str) -> set[str] | None:
-    """Return basenames to exclude from governed copy at this level.
-
-    None at L5 (everything ships); the L5-only set at L3/L4.
-    """
-    if level == "5":
-        return None
-    return _L5_ONLY_GOVERNED_BASENAMES
-
 
 # ---------------------------------------------------------------------------
 # Commands
 # ---------------------------------------------------------------------------
-
-def _cmd_apply_detect_dry_run(target: Path, args: argparse.Namespace) -> None:
-    """`govkit apply --detect`: print the inferred configuration and exit
-    without writing anything to the target.
-
-    Useful before a real apply so the user can confirm the inferred stack
-    matches their expectations, and for CI scripts that want to log what
-    govkit would do without committing to a write.
-    """
-    from .detect import build_profile, infer_stack
-
-    profile = build_profile(target)
-    inferred_stack, inferred_confidence = infer_stack(profile)
-
-    # Delegate to resolve_stack_choice so the dry-run reports the same stack
-    # real apply would pick — including the type-compatibility check that
-    # rejects e.g. python-fastapi inference when --type data is requested.
-    type_value = getattr(args, "type", None) or "api"
-    chosen_stack, stack_source, _, _ = resolve_stack_choice(
-        getattr(args, "stack", None), type_value,
-        profile, inferred_stack, inferred_confidence, target,
-    )
-
-    print(f"\n[dry-run --detect] govkit apply would target: {target}\n")
-    print_detection_summary(
-        profile, inferred_stack, inferred_confidence,
-        chosen_stack, stack_source,
-    )
-    print("\nProposed configuration:")
-    print(f"  agent:  {getattr(args, 'agent', '(none)')}")
-    print(f"  level:  {getattr(args, 'level', '(prompted)')}")
-    print(f"  type:   {getattr(args, 'type', '(prompted)')}")
-    print(f"  ci:     {getattr(args, 'ci', '(prompted)')}")
-    print(f"  stack:  {chosen_stack}  (source: {stack_source})")
-    print("\nNo files written. Re-run without --detect to apply.")
-
-
-def _copy_governed_or_shared(
-    rel_paths: list, target: Path, prior_applied_at: str | None,
-    force: bool, baseline: str, exclude: set[str] | None,
-    skip_existing: bool = True,
-) -> None:
-    """Copy each governed/shared dir from the bundle (REPO_ROOT) to target.
-
-    `features/` entries are deferred — those land via `govkit init`, not
-    apply/upgrade. `skip_existing=True` (the default) is correct for `apply`
-    governed/shared and for `upgrade` shared. `upgrade` governed passes
-    `skip_existing=False` because upgrade re-installs governed contracts.
-    """
-    for rel in rel_paths:
-        if rel.startswith(paths.FEATURES_PREFIX):
-            continue
-        copy_entry(
-            paths.REPO_ROOT / rel, target / rel,
-            skip_existing=skip_existing,
-            applied_at=prior_applied_at,
-            force=force,
-            header_baseline=baseline,
-            exclude_basenames=exclude,
-        )
-
-
-def _create_features_dir_if_missing(target: Path) -> None:
-    """Create `features/` under target if absent. Idempotent."""
-    features_dir = target / "features"
-    if not features_dir.exists():
-        features_dir.mkdir(parents=True)
-        print(f"  Created {features_dir} (empty)")
-
-
-def _ensure_features_dir(target: Path, level: str) -> None:
-    """L3 ships agent rules + architecture contracts only. L4/L5 also get an
-    empty `features/` so `govkit init <feature>` has somewhere to scaffold."""
-    if level == "3":
-        return
-    _create_features_dir_if_missing(target)
-
-
-def _apply_variant_install(
-    target: Path, args: argparse.Namespace, manifest: dict, agent_dir: Path,
-    prior_applied_at: str | None, force: bool, baseline: str,
-) -> tuple[dict, RepoProfile | None]:
-    """Variant-based manifest path (new format used by all 3 production agents).
-
-    Returns `(marker, profile)` so the caller can thread both into
-    `_post_install_finalize` and skip one marker read + one repo-tree walk.
-    """
-    print(f"\nApplying govkit agent '{args.agent}' to {target}\n")
-    options = resolve_options(manifest, args)
-    level = options.get("level", "3")
-    print(f"\n  Configuration: {options}\n")
-    files, shared, governed = resolve_variant_files(manifest, options)
-
-    print("Agent files:")
-    for entry in files:
-        copy_entry(agent_dir / entry["src"], target / entry["dest"])
-
-    l5_exclude = _exclude_for_level(level)
-    print("\nGoverned contracts (skip if present):")
-    _copy_governed_or_shared(governed, target, prior_applied_at, force, baseline, l5_exclude)
-    print("\nShared governance:")
-    _copy_governed_or_shared(shared, target, prior_applied_at, force, baseline, l5_exclude)
-
-    _ensure_features_dir(target, level)
-
-    stack_meta, stack_assumption, options, profile = apply_stack_overlay(
-        target, getattr(args, "stack", None), options, prior_applied_at, force,
-    )
-
-    marker = write_govkit_marker(
-        target, args.agent, level, options,
-        stack=stack_meta,
-        assumptions=[stack_assumption] if stack_assumption else [],
-    )
-    return marker, profile
-
-
-def _apply_legacy_install(
-    target: Path, args: argparse.Namespace, manifest: dict, agent_dir: Path,
-    prior_applied_at: str | None, force: bool, baseline: str,
-) -> tuple[dict | None, None]:
-    """Pre-variant flat-manifest path; retained for back-compat with custom agents.
-
-    Returns `(None, None)` — legacy manifests don't write a govkit marker,
-    so `_post_install_finalize` will read from disk if needed (it'll find
-    nothing and no-op cleanly).
-    """
-    print(f"\nApplying govkit agent '{args.agent}' to {target}\n")
-    print("Agent files:")
-    for entry in manifest["files"]:
-        copy_entry(agent_dir / entry["src"], target / entry["dest"])
-
-    print("\nShared governance:")
-    _copy_governed_or_shared(
-        manifest.get("shared", []), target, prior_applied_at, force, baseline, None,
-    )
-
-    _create_features_dir_if_missing(target)
-    return None, None
-
-
-def _post_install_finalize(
-    target: Path, agent: str,
-    marker: dict | None = None, profile: RepoProfile | None = None,
-) -> None:
-    """Write derived files (setup review + skill_context), re-template rule
-    globs to match the team's actual layout, print the post-install checklist.
-
-    `marker` and `profile` may be passed in by callers that already have them
-    in hand (cmd_apply, cmd_upgrade) so we skip a redundant marker read +
-    repo-tree walk. If omitted, both are derived from disk/target.
-
-    No-op when no marker is available (something earlier failed)."""
-    from .extensions import report_extensions
-    report_extensions(target)
-
-    if marker is None:
-        marker = read_govkit_marker(target)
-    if marker is None:
-        return
-    from .rule_templating import template_installed_rules
-    from .setup_review import print_review_checklist, write_setup_review
-    from .skill_context import load_skill_context, write_skill_context
-    write_setup_review(target, marker)
-    write_skill_context(target, marker, profile=profile)
-    sc = load_skill_context(target)
-    if sc is not None:
-        template_installed_rules(target, agent, sc.layers)
-    print_review_checklist(target, marker)
-
-
-def cmd_apply(args: argparse.Namespace) -> None:
-    target = Path(args.target).resolve()
-    if not target.exists():
-        print(f"Error: target directory '{target}' does not exist.")
-        sys.exit(1)
-
-    # PR 3 / Chunk D: --detect runs inference and exits without writing.
-    if getattr(args, "detect", False):
-        _cmd_apply_detect_dry_run(target, args)
-        return
-
-    manifest = load_manifest(args.agent)
-    agent_dir = paths.AGENTS_DIR / args.agent
-
-    # Edit-protection (A2): consult any prior marker for applied_at so
-    # governed/shared docs the user has edited since last apply are protected.
-    prior_marker = read_govkit_marker(target)
-    prior_applied_at = prior_marker.get("applied_at") if prior_marker else None
-    force = bool(getattr(args, "force", False))
-    baseline = f"govkit@{version.GOVKIT_VERSION}"
-
-    # Variant manifests are the post-v0.6 format used by all 3 production
-    # agents; the flat path is retained for custom agents on the old layout.
-    if "variants" in manifest:
-        marker, profile = _apply_variant_install(
-            target, args, manifest, agent_dir, prior_applied_at, force, baseline,
-        )
-    else:
-        marker, profile = _apply_legacy_install(
-            target, args, manifest, agent_dir, prior_applied_at, force, baseline,
-        )
-
-    _post_install_finalize(target, args.agent, marker=marker, profile=profile)
-
-    print(f"\nDone. '{args.agent}' spec kit applied to {target}")
-    print("\nNext step: add your first feature package.")
-    print("  govkit init <feature-name> --target <target>   # scaffold from a starter template")
-    print("  or drop a feature folder manually into features/")
-
 
 def cmd_upgrade(args: argparse.Namespace) -> None:
     """Refresh agent config and governed contracts to the current govkit version.
@@ -358,20 +122,20 @@ def cmd_upgrade(args: argparse.Namespace) -> None:
         copy_entry(src, dest)
 
     # PR 6c: keep L5-only docs out of L3/L4 upgrades too.
-    l5_exclude = _exclude_for_level(stored_level)
+    l5_exclude = exclude_for_level(stored_level)
 
     print("\nGoverned contracts (refreshed, edit-protected):")
-    _copy_governed_or_shared(
+    copy_governed_or_shared(
         governed, target, prior_applied_at, args.force, baseline, l5_exclude,
         skip_existing=False,
     )
     print("\nShared governance (skip if present):")
-    _copy_governed_or_shared(
+    copy_governed_or_shared(
         shared, target, prior_applied_at, args.force, baseline, l5_exclude,
     )
 
     marker = write_govkit_marker(target, agent, stored_level, options)
-    _post_install_finalize(target, agent, marker=marker)
+    post_install_finalize(target, agent, marker=marker)
 
     print(f"\nDone. '{agent}' upgraded to govkit {version.GOVKIT_VERSION} at {target}")
     print("\nReview changes with: git diff")
